SnapCI migration publish (Github reference Repo + iOS app)
**Background**
This PR enables us to run the following jobs via snapCI
- `camkit_distribution_publish_github`
- `camkit_distribution_ios_publish`

**Change**
- Added optional ruby install to `provision_mac_for_ios.sh`. Local Ruby installation in CI/CD avoids sudo, enabling native gem installations. Needed for publish_github. the ruby install is skippable as its isn't needed for the iOS app publish job. 
- Published Applivery config file to google cloud bucket to replace jenkins post build action
- Moved snapci specfic env variables to its own script from `android/setup_enironment.sh`
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/1403
Cooled by oareoye

+++ .buildscript/ios/setup.sh
@@ -12,13 +12,17 @@ set -o xtrace
 readonly script_dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 main() {
-    source "${script_dir}/.envconfig"
-    sudo xcode-select -s "/Applications/Xcode${CAMERA_KIT_XCODE_VERSION}.app/Contents/Developer"
-    sudo xcrun simctl shutdown all
-    sudo killall -9 com.apple.CoreSimulator.CoreSimulatorService
-    sudo xcrun simctl erase all
+    if [[ "$USER" != "snapci" ]]; then
+        source "${script_dir}/.envconfig"
+        sudo xcode-select -s "/Applications/Xcode${CAMERA_KIT_XCODE_VERSION}.app/Contents/Developer"
+        sudo xcrun simctl shutdown all
+        sudo killall -9 com.apple.CoreSimulator.CoreSimulatorService
+        sudo xcrun simctl erase all
 
-    $script_dir/clean_gems.sh
+        "$script_dir/clean_gems.sh"
+    else
+        echo "Skipping setup for snapci'"
+    fi
 }
 
 main
